# backend/docs/openapi.yml
openapi: 3.0.3
info:
  title: Campus Shop Admin API
  version: 1.0.0
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, example: admin@campus.edu }
        password: { type: string, example: admin123 }
    LoginResponse:
      type: object
      properties:
        accessToken: { type: string }
        user:
          type: object
          properties:
            id: { type: integer }
            email: { type: string }
            role: { type: string, example: admin }
    User:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string }
        status: { type: string, example: active }
        warnings: { type: integer, example: 0 }
    Listing:
      type: object
      properties:
        id: { type: integer }
        title: { type: string }
        status: { type: string, example: pending }
security:
  - bearerAuth: []
paths:
  /api/auth/login:
    post:
      summary: Login to get JWT
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/LoginRequest' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LoginResponse' }
  /api/admin/users:
    get:
      summary: List users (exclude admin)
      security: [ { bearerAuth: [] } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/User' }
  /api/admin/users/{id}/warn:
    post:
      summary: Warn a user
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                message: { type: string, example: Please follow community guidelines. }
      responses:
        '200': { description: Warned }
  /api/admin/users/{id}/suspend:
    post:
      summary: Suspend a user
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, example: Policy violation }
      responses:
        '200': { description: Suspended }
  /api/admin/listings:
    get:
      summary: List all listings
      security: [ { bearerAuth: [] } ]
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Listing' }
  /api/admin/listings/{id}/approve:
    post:
      summary: Approve a listing
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Approved }
  /api/admin/listings/{id}:
    put:
      summary: Edit a listing (title/desc/price…)
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200': { description: Updated }
    delete:
      summary: Remove a listing
      security: [ { bearerAuth: [] } ]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Removed }